BlackRock Assignment/backend/simulation.py

from flask import Flask, request, jsonify
from asgiref.wsgi import WsgiToAsgi
# Initialize the Flask application
app = Flask(__name__)

# Define the simulation function that models the Lorenz system
def simulate(x0, y0, z0, sigma, rho, beta, delta_t, n):
    '''
    Simulate the Lorenz equations over 'n' time steps.

    :param x0: Initial x-coordinate
    :param y0: Initial y-coordinate
    :param z0: Initial z-coordinate
    :param sigma: Lorenz system parameter
    :param rho: Lorenz system parameter
    :param beta: Lorenz system parameter
    :param delta_t: Time step size
    :param n: Number of time steps to simulate
    :return: List of dictionaries containing the system state at each time step
    '''
    # Initialize the results list
    results = []
    # Set the initial state of the system
    x, y, z = x0, y0, z0
    # Iterate over 'n' time steps
    for i in range(n):
        # Calculate the next state using the Lorenz equations
        x_next = x + sigma * (y - x) * delta_t
        y_next = y + (x * (rho - z) - y) * delta_t
        z_next = z + (x * y - beta * z) * delta_t
        # Update the state
        x, y, z = x_next, y_next, z_next
        # Append the new state to the results list
        results.append({'n': i, 'x': x, 'y': y, 'z': z})
    # Return the list of results
    return results

# Define the route for the simulation API endpoint
@app.route('/simulate', methods=['POST'])
def run_simulation_api():
    '''
    Handle POST request to run the Lorenz system simulation.

    Expects JSON payload with initial conditions and parameters.
    Returns JSON response with simulation results.
    '''
    # Parse the JSON payload from the request
    data = request.get_json()
    # Extract parameters from the payload
    x0 = data.get('x0')
    y0 = data.get('y0')
    z0 = data.get('z0')
    sigma = data.get('sigma')
    rho = data.get('rho')
    beta = data.get('beta')
    delta_t = data.get('delta_t')
    n = data.get('n')
    
    # Validate that none of the required parameters are missing
    if not all([x0 is not None, y0 is not None, z0 is not None, sigma is not None, 
                rho is not None, beta is not None, delta_t is not None, n is not None]):
        # If any are missing, return an error response
        return jsonify({"error": "Missing data for simulation"}), 400
    
    # If all parameters are present, run the simulation
    simulation_results = simulate(x0, y0, z0, sigma, rho, beta, delta_t, n)
    
    # Return the simulation results as a JSON response
    return jsonify({'results': simulation_results})

# Check if the run is the main application run
if __name__ == '__main__':
    # Start the Flask application with debugging enabled
    app.run(debug=True)
    haystack = ""
    needle = ""
    haystack.find(needle)

app = WsgiToAsgi(app)